+++ HISTORY.md
@@ -1,7 +1,12 @@
+0.1.12 / 30.11.2016
+===================
+
+  * Managed to reduce metadata size by another 5 KiloBytes removing redundant (duplicate) phone number type regular expressions (because there's no "get phone type" API in this library).
+
 0.1.11 / 30.11.2016
 ===================
 
-  * Managed to reduce metadata by 10 KiloBytes removing phone number type regular expressions when `leading_digits` are present.
+  * Managed to reduce metadata size by 10 KiloBytes removing phone number type regular expressions when `leading_digits` are present.
 
 0.1.10 / 30.11.2016
 ===================